% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHtests_merge.R
\name{filter_TP}
\alias{filter_TP}
\title{merge monthly and yearly TPs and filter out bad ones}
\usage{
filter_TP(d, nyear = 1)
}
\arguments{
\item{d}{A data.frame returned by \code{\link[=tidy_TP_sites]{tidy_TP_sites()}}}

\item{nyear}{the monthly TP will be filter out, if the distance to the
nearest yearly TP is longer than \code{nyear}.}
}
\description{
merge monthly and yearly TPs and filter out bad ones
}
