% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-find_refer.R
\name{find_refer}
\alias{find_refer}
\title{select reference sites}
\usage{
find_refer(df2, varname = "Tavg", sites_worst)
}
\arguments{
\item{df2}{a data.frame with columns of \code{site}, \code{date}, and \code{varname}}

\item{sites_worst}{sites with significant TPs}
}
\description{
select reference sites
}
