% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_uncontinue.R
\name{fix_uncontinue}
\alias{fix_uncontinue}
\title{fix un-continuous observations for the INPUT data}
\usage{
fix_uncontinue(df, complete_year = TRUE)
}
\arguments{
\item{df}{A data.frame with the columns of \code{site}, \code{date} and others}

\item{complete_year}{Boolean. Note that \code{complete_year} has to be true for RHtests.}
}
\description{
fix un-continuous observations for the INPUT data and only keep the complete
year.
}
