% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-revise_locaiton_multi.R
\name{get_dist}
\alias{get_dist}
\title{The distance of locations from the location having the}
\usage{
get_dist(lon, lat, n_period)
}
\description{
The distance of locations from the location having the
}
