% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHtests_In.R
\name{getInput_refer}
\alias{getInput_refer}
\title{getInput_refer}
\usage{
getInput_refer(df_mat, st_refer)
}
\arguments{
\item{df_mat}{A data.table, with the daily data and the columns of \code{c(date, sites)}}

\item{st_refer}{\itemize{
\item \code{target}:
\item \code{site}  : reference site
\item \code{cor}   : correlation of target and reference site
}}
}
\description{
getInput_refer
}
