% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homo_ref.R
\name{homo_ref}
\alias{homo_ref}
\title{Homogeneous time series data with reference series}
\usage{
homo_ref(d, metedata = NULL)
}
\arguments{
\item{d}{A data.frame with columns of date, varname, ref}

\item{metedata}{A data.frame with column date indicating turning point}
}
\description{
Homogeneous time series data with reference series
}
