% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenize.wRef.R
\name{homogenize.wRef.list}
\alias{homogenize.wRef.list}
\title{get the turning points of monthly and yearly data}
\usage{
homogenize.wRef.list(lst, st_moveInfo, .parallel = FALSE, .debug = FALSE)
}
\arguments{
\item{lst}{List of multiple sites data.frame
The data.frame of each site:
\itemize{
\item if 3 columns, they are \code{date}, \code{varname} and \code{ref}, and homo with ref will
be used.
\item if 2 columns, they are \code{date}, \code{varname}, and homo without ref will
be used.
}}

\item{st_moveInfo}{A data.frame with the site relocation moving information}
}
\description{
get the turning points of monthly and yearly data
}
\seealso{
\code{\link[=homogenize.wRef]{homogenize.wRef()}}
}
