% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenize.wRef.R
\name{homogenize.wRef}
\alias{homogenize.wRef}
\title{Homogeneous time series data with reference series}
\usage{
homogenize.wRef(d, metadata = NULL, prefix = "./OUTPUT/temp/example01")
}
\arguments{
\item{d}{A data.frame of the inputdata:
\itemize{
\item if 3 columns, they are \code{date}, \code{varname} and \code{ref}, and homo with ref will
be used.
\item if 2 columns, they are \code{date}, \code{varname}, and homo without ref will
be used.
}}

\item{metedata}{A data.frame with column date indicating turning point}
}
\description{
Homogeneous time series data with reference series
}
