% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenize.R
\name{homogenize_monthly}
\alias{homogenize_monthly}
\title{RHtests_main}
\usage{
homogenize_monthly(df, st_moveInfo, sites, varname)
}
\arguments{
\item{df}{A data.frame with the columns at least of \code{site}, 'date', 'varname'}

\item{st_moveInfo}{}
}
\description{
RHtests_main
}
