% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_mete2000_txts.R
\name{merge_mete2000_txts}
\alias{merge_mete2000_txts}
\title{merge txt files into a csv}
\usage{
merge_mete2000_txts(dir_root, is_save = TRUE)
}
\arguments{
\item{dir_root}{e.g. \code{dir_root = "N:/DATA/China/2400climate data" \%>\% path.mnt()}}
}
\description{
merge txt files into a csv
}
\details{
\subsection{WIN errors fixed:}{

\link{826} 69473, 59034 2429 10839   1851 2019 10 235000002     64      5    106      5 9 9 9 9 9
win_avg: 5000002 -> 20
\link{827} 17129: 53955 3531 11028   4616 2019 11 292500012     33      3     64      3 9 9 9 9 9
win_avg: 500012 -> 12
\link{829}  3792: 51238 4454  8204   5322 2020  1  72500001     21     13     32     13 9 9 9 9 9
win_avg: 2500001 -> 10
}
}
\examples{
\dontrun{
dir_root = "N:/DATA/China/2400climate data" \%>\% path.mnt()
merge_mete2000_txts(dir_root, is_save = TRUE)
}
}
