% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RHtests.R
\name{merge_refer2}
\alias{merge_refer2}
\title{merge_refer2}
\usage{
merge_refer2(df, l_Ref_day, l_noRef_day, ...)
}
\arguments{
\item{df}{with the column of \code{c("site", "date", "value")}}
}
\description{
merge_refer2
}
