% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_refer.R
\name{period_coverage}
\alias{period_coverage}
\title{the percentage of overlaped date period of x and y to the length of period x}
\usage{
period_coverage(x, y)
}
\arguments{
\item{x}{A vector with \code{date_begin} and \code{date_end}.}

\item{y}{A data.frame with the columns of \code{date_begin} and \code{date_end}}
}
\description{
the percentage of overlaped date period of x and y to the length of period x
}
