% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{RHtests process}
\usage{
process(df = NULL, varname = "Tavg", overwrite = FALSE)
}
\arguments{
\item{df}{A data.table with the columns of \code{site}, \code{lon}, \code{lat}, \code{alt}, \code{date}
and \code{varname}.}

\item{varname}{String, the variable that you wish to harmonized.}

\item{overwrite}{Boolean}
}
\description{
RHtests process
}
