% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RHtests.R
\name{query_siteHomoInfo}
\alias{query_siteHomoInfo}
\title{query_siteHomoInfo}
\usage{
query_siteHomoInfo(res, TP = NULL)
}
\value{
\itemize{
\item \code{Yes}: has no TP
\item \code{No} : high confident TP
\item \verb{?}  ：low confident TP
}
}
\description{
query_siteHomoInfo
}
