% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_Ipaper.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{colMeans2}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{matrixStats}{\code{\link[matrixStats:rowMeans2]{colMeans2}}}
}}

