% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revise_locaiton_multi.R
\name{revise_locaiton}
\alias{revise_locaiton}
\title{Revise meteorological station's location error}
\usage{
revise_locaiton(
  d,
  prefix = "",
  dist_max = 100,
  score_min = 2,
  vars_rm = c("date_begin", "date_end", "n_all"),
  verbose = 2
)
}
\arguments{
\item{d}{A data.frame with the columns at least of \code{dist}, \code{lon}, \code{lat}, \code{alt}
and \code{n_period}, \code{QC}}

\item{score_min}{if distance score lower than \code{score_min}, will not fix}

\item{vars_rm}{variables not shown in the console}
}
\value{
sites with not outliers, QC = "";

sites with not outliers, for outliers, QC equals:
score = 0: unfixed
score = 1: marginal quality fixed
score = 2: good quality fixed
}
\description{
Revise meteorological station's location error
}
