% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revise_locaiton_multi.R
\name{revise_locaiton_multi}
\alias{revise_locaiton_multi}
\title{Revise site moving information for all sites}
\usage{
revise_locaiton_multi(info, dist_max = 50)
}
\arguments{
\item{info}{site moving info returned by \code{\link[=get_moveInfo]{get_moveInfo()}}}

\item{dist_max}{If the site moving distance beyond the \code{dist_max}, it will
be regarded as outlier and will be fixed by \code{\link[=revise_location]{revise_location()}}.}
}
\description{
Revise site moving information for all sites
}
\details{
Continuous moving location outliers are also considered at here.
}
