% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHtests.R
\name{RHtests_main}
\alias{RHtests_main}
\title{RHtests_main}
\usage{
RHtests_main(df, st_moveInfo, sites, varname)
}
\arguments{
\item{df}{A data.frame with the columns at least of \code{site}, 'date', 'varname'}

\item{st_moveInfo}{}
}
\description{
RHtests_main
}
