% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHtestsHelper-package.R
\docType{package}
\name{RHtestsHelper-package}
\alias{RHtestsHelper}
\alias{RHtestsHelper-package}
\title{RHtestsHelper: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
