% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHtests_merge.R
\name{tidy_TP_sites}
\alias{tidy_TP_sites}
\alias{tidy_TP_site}
\title{tidy_TP_sites}
\usage{
tidy_TP_sites(res2)

tidy_TP_site(x)
}
\arguments{
\item{res2}{object returned by \code{\link[=RHtests]{RHtests()}}}
}
\description{
tidy_TP_sites
}
