% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RHtests.R
\name{TP_highConf}
\alias{TP_highConf}
\title{merge monthly and yearly TPs and mask bads out}
\usage{
TP_highConf(d, nyear = 1)
}
\arguments{
\item{d}{A data.frame returned by \code{\link[=TP_mergeYM_sites]{TP_mergeYM_sites()}}}

\item{nyear}{the monthly TP will be filter out, if the distance to the

nearest yearly TP is longer than \code{nyear}.}
}
\description{
merge monthly and yearly TPs and mask bads out
}
