% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP_merge.R
\name{TP_mask}
\alias{TP_mask}
\title{merge monthly and yearly TPs and mask bads out}
\usage{
TP_mask(d, nyear = 1)
}
\arguments{
\item{d}{A data.frame returned by \code{\link[=TP_tidy_sites]{TP_tidy_sites()}}}

\item{nyear}{the monthly TP will be filter out, if the distance to the
nearest yearly TP is longer than \code{nyear}.}
}
\description{
merge monthly and yearly TPs and mask bads out
}
