% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP_merge.R
\name{TP_mergeYM}
\alias{TP_mergeYM}
\alias{TP_mergeYM_sites}
\title{TP_mergeYM_sites}
\usage{
TP_mergeYM(l, site = NULL)

TP_mergeYM_sites(res)
}
\arguments{
\item{l}{Object returned by \code{\link[=homogenize.wRef]{homogenize.wRef()}}, a list with the elements of
\itemize{
\item year
\item month
}}

\item{res}{object returned by \code{\link[=RHtests]{RHtests()}}}
}
\description{
TP_mergeYM_sites
}
