% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP_merge.R
\name{TP_tidy_sites}
\alias{TP_tidy_sites}
\alias{TP_tidy_site}
\title{TP_tidy_sites}
\usage{
TP_tidy_sites(res2)

TP_tidy_site(x)
}
\arguments{
\item{res2}{object returned by \code{\link[=RHtests]{RHtests()}}}

\item{x}{Object returned by \code{\link[=homo_ref]{homo_ref()}}, a list with the elements of
\itemize{
\item year
\item month
\item day
}}
}
\description{
TP_tidy_sites
}
