% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenize.wRef.R
\name{homogenize.wRef}
\alias{homogenize.wRef}
\alias{homo_withRef}
\alias{homo_withRef_multi}
\title{Homogeneous time series data with reference series}
\usage{
homogenize.wRef(
  d,
  metadata = NULL,
  prefix = "./OUTPUT/example01",
  verbose = FALSE
)

homo_withRef(
  site_target,
  df_day,
  st_refs,
  siteHomoInfo,
  nmax = 5,
  ...,
  debug = TRUE
)

homo_withRef_multi(
  df_day,
  st_refs,
  siteHomoInfo,
  nmax = 5,
  ...,
  debug = FALSE,
  .parallel = TRUE
)
}
\arguments{
\item{d}{A data.frame of the inputdata:
\itemize{
\item if 3 columns, they are \code{date}, \code{varname} and \code{ref}, and homo with ref will
be used.
\item if 2 columns, they are \code{date}, \code{varname}, and homo without ref will
be used.
}}

\item{metedata}{A data.frame with column date indicating turning point}
}
\description{
Get the turning points of monthly and yearly data
}
