% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHtests_main.R
\name{RHtests_main}
\alias{RHtests_main}
\title{get the turning points of monthly and yearly data}
\usage{
RHtests_main(df, st_moveInfo, sites, varname, .parallel = FALSE)
}
\arguments{
\item{df}{A data.frame with the columns at least of \code{site}, 'date', 'varname'}

\item{st_moveInfo}{}
}
\description{
get the turning points of monthly and yearly data
}
